pyproject.toml
[project]
name = "voicereadertg"
version = "0.1.0"
description = "Telegram bot that summarizes documents with voice notes and offers VAPI-powered voice discussions"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-telegram-bot[ext]>=21.7",
    "anthropic>=0.40.0",
    "openai>=1.55.0",
    "pdfplumber>=0.11.0",
    "python-docx>=1.1.0",
    "trafilatura>=2.0.0",
    "pydantic-settings>=2.6.0",
    "jinja2>=3.1.0",
    "httpx>=0.28.0",
    "python-multipart>=0.0.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
